refactor(all) : validator 클래스 삭제, validate 함수 필요한 클래스에 추가
