+++ src/main/java/vendingmachine/domain/Merchandises.java
@@ -3,24 +3,25 @@ package vendingmachine.domain;
 import static vendingmachine.domain.Merchandise.*;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import vendingmachine.utils.ErrorMessage;
-import vendingmachine.utils.Validator;
 
 public class Merchandises {
 	private List<Merchandise> merchandiseList;
 
 	public Merchandises(List<Merchandise> merchandiseList) {
-		Validator.validateDuplicateMerchandise(merchandiseList);
+		validateDuplicateMerchandise(merchandiseList);
 		this.merchandiseList = merchandiseList;
 	}
 
 	public static List<Merchandise> constructMerchandises(String merchandiseInformations) {
 		List<Merchandise> merchandiseList = new ArrayList<>();
 		for (String merchandiseInformation : merchandiseInformations.split(MERCHANDISE_PARSER)) {
-			Validator.validateInputMerchandise(merchandiseInformation);
+			validateInputMerchandise(merchandiseInformation);
 			String merchandise = merchandiseInformation.substring(1, merchandiseInformation.length() - 1).trim();
 			merchandiseList.add(constructMerchandise(merchandise));
 		}
@@ -52,5 +53,18 @@ public class Merchandises {
 
 	public int getMerchandisesSize() {
 		return merchandiseList.size();
+	}
+
+	public static void validateDuplicateMerchandise(List<Merchandise> merchandiseList) {
+		Set<Merchandise> merchandiseSet = new HashSet<>(merchandiseList);
+		if (merchandiseSet.size() != merchandiseList.size()) {
+			throw new IllegalArgumentException(ErrorMessage.INVALID_DUPLICATED_MERCHANDISE_ERROR_MESSAGE);
+		}
+	}
+
+	public static void validateInputMerchandise(String merchandiseInformation) {
+		if (!merchandiseInformation.contains("[") || !merchandiseInformation.contains("]")) {
+			throw new IllegalArgumentException(ErrorMessage.INVALID_MERCHANDISE_INPUT_ERROR_MESSAGE);
+		}
 	}
 }
