+++ src/main/java/vendingmachine/domain/Money.java
@@ -1,13 +1,13 @@
 package vendingmachine.domain;
 
-import vendingmachine.utils.Validator;
+import vendingmachine.utils.ErrorMessage;
 
 public class Money {
 	private int money;
 
 	public Money(int money) {
-		Validator.validateMoney(money);
-		Validator.validateDivideMoneyBy10Coin(money);
+		validateMoney(money);
+		validateDivideMoneyBy10Coin(money);
 		this.money = money;
 	}
 
@@ -24,5 +24,17 @@ public class Money {
 			return true;
 		}
 		return false;
+	}
+
+	public static void validateMoney(int money) {
+		if (money <= 0) {
+			throw new IllegalArgumentException(ErrorMessage.INVALID_MONEY_ERROR_MESSAGE);
+		}
+	}
+
+	public static void validateDivideMoneyBy10Coin(int merchandisePrice) {
+		if (merchandisePrice % 10 != 0) {
+			throw new IllegalArgumentException(ErrorMessage.INVALID_MERCHANDISE_PRICE_NOT_DIVIDE_10_COIN_ERROR_MESSAGE);
+		}
 	}
 }
