+++ src/main/java/vendingmachine/domain/Name.java
@@ -1,16 +1,22 @@
 package vendingmachine.domain;
 
-import vendingmachine.utils.Validator;
+import vendingmachine.utils.ErrorMessage;
 
 public class Name {
 	private final String name;
 
 	public Name(String name) {
-		Validator.validateEmptyMerchandiseName(name);
+		validateEmptyMerchandiseName(name);
 		this.name = name;
 	}
 
 	public boolean isSameName(String anotherName) {
 		return name.equals(anotherName);
+	}
+
+	public static void validateEmptyMerchandiseName(String merchandiseName) {
+		if (merchandiseName.isEmpty()) {
+			throw new IllegalArgumentException(ErrorMessage.INVALID_MERCHANDISE_EMPTY_NAME_ERROR_MESSAGE);
+		}
 	}
 }
