+++ src/main/java/vendingmachine/domain/Quantity.java
@@ -1,12 +1,12 @@
 package vendingmachine.domain;
 
-import vendingmachine.utils.Validator;
+import vendingmachine.utils.ErrorMessage;
 
 public class Quantity {
 	private int quantity;
 
 	public Quantity(int quantity) {
-		Validator.validateQuantity(quantity);
+		validateQuantity(quantity);
 		this.quantity = quantity;
 	}
 
@@ -16,5 +16,11 @@ public class Quantity {
 
 	public boolean isQuantityZero() {
 		return quantity == 0;
+	}
+
+	public static void validateQuantity(int quantity) {
+		if (quantity < 0) {
+			throw new IllegalArgumentException(ErrorMessage.INVALID_QUANTITY_ERROR_MESSAGE);
+		}
 	}
 }
